220429 mypage 수정2
