--- qCali/src/main/java/com/group/exam/member/command/MemberFindPwdCommand.java
@@ -1,21 +1,0 @@
-package com.group.exam.member.command;
-
-import org.hibernate.validator.constraints.Email;
-import org.hibernate.validator.constraints.NotEmpty;
-
-public class MemberFindPwdCommand {
-	@Email(message = "이메일 형식이 아닙니다.")
-	@NotEmpty(message = "이메일을 입력하세요.")
-	private String memberId;
-
-	public String getMemberId() {
-		return memberId;
-	}
-
-	public void setMemberId(String memberId) {
-		this.memberId = memberId;
-	}
-
-
-
-}
