+++ qCali/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -66,8 +66,8 @@
 		<beans:property name="host" value="smtp.gmail.com" />
 		<beans:property name="port" value="587" />
 		<beans:property name="username"
-			value="ejdd9630@gmail.com" />
-		<beans:property name="password" value="****" />
+			value="구글계정 아이디" />
+		<beans:property name="password" value="구글계정 비밀번호" />
 		<beans:property name="javaMailProperties">
 			<beans:props>
 				<beans:prop key="mail.transport.protocol">smtp</beans:prop>
