--- qCali/src/main/webapp/WEB-INF/views/errors/mypageChangeError.jsp
@@ -1,7 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-    <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
-<script>
-	alert("회원 정보 수정 에러! \n 메인 화면으로 돌아갑니다.");
-	document.location.href="<c:url value='/'/>";
-</script>
