+++ qCali/src/main/webapp/WEB-INF/views/member/findPwdForm.jsp
@@ -1,10 +1,12 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
+	pageEncoding="UTF-8"%>
 <%@ taglib prefix="form" uri="http://www.springframework.org/tags/form"%>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
 <!DOCTYPE html>
 <html>
 <head>
+<script src="//cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+<script src="https://code.jquery.com/jquery-3.6.0.js"></script>
 <style>
 body {
 	text-align: center;
@@ -20,22 +22,46 @@ table {
 <title>비밀번호 찾기</title>
 </head>
 <body>
-	<h2>비밀번호 찾기</h2>
-	<form:form commandName="findPwdData">
-		<table border="1">
-			<tr>
-				<td>아이디 : </td>
-				<td><form:input path="memberId" placeholder="필수입력"/>
-				<form:errors path="memberId"/></td>
-			</tr>
-		
-		</table>
-		<br>
-		${msg}<br><br>
-		<input type="submit" value="비밀번호 찾기" />
-	</form:form>
+	<h2>가입시 사용한 이메일을 입력해주세요.</h2>
 
-		
+
+	<input type="text" id="memberId" />
+
+	<br>
+
+	<button type="button" onClick="submit_close()">비밀번호 찾기</button>
+
+
+	<script>
+		function submit_close() {
+
+			let memberId = $('#memberId').val();
+
+			$.ajax({
+				url : '${pageContext.request.contextPath}/member/findPwd',
+				type : 'POST',
+				data : memberId,
+				contentType :'application/json',
+				success : function(result) {
+					console.log(result);
+					console.log(result == false);
+					console.log(result == true);
+					
+					
+					if(result == true){
+					alert("이메일로 임시 비밀번호가 발송되었습니다.");
+					window.close();
+					
+					} else {
+						
+						alert("회원 정보가 없습니다.");
+						
+					}
+				},
+			});
+		}
+	</script>
+
 
 </body>
 </html>
