--- qCali/src/main/webapp/WEB-INF/views/member/member_alert/findPwdNext.jsp
@@ -1,7 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-    <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
-<script>
-	alert("해당 이메일로 임시 비밀번호를 발송했습니다.");
-	document.location.href="<c:url value='/member/login'/>";
-</script>
