--- qCali/src/main/webapp/WEB-INF/views/member/member_alert/signUpSuccess.jsp
@@ -1,14 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>회원가입 완료</title>
-</head>
-<body>
-이메일 인증 성공! 로그인 화면으로 이동합니다.<br><br>
-<a href="${pageContext.request.contextPath }/member/login">로그인 하기</a>
-
-</body>
-</html>
